started working on support for private conversations
I'm just committing the dirty working directory that I found from months ago,
because I'm never going to finish this project. I have no idea if it's good.
Conversation support doesn't seem to be finished, if my testing is any metric.
But it's a fair amount of code and it seems a waste to throw it away.